Load JuiceVlk.dll for vkGetInstanceProcAddress using user DLL search dirs.
The directory to load JuiceVlk.dll from is set by AddDllDirectory() in
JuiceShim{32,64}.dll but this search path is only used in calls to
LoadLibraryEx() that pass the LOAD_LIBRARY_SEARCH_USER_DIRS flag.  The
LOAD_LIBRARY_SEARCH_DEFAULT_DIRS flag that is used also searches the
system directory for dbghelp.dll which is loaded by JuiceVlk.dll.